requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/dev.txt requirements/dev.in
#
astroid==1.5.3            # via pylint
autoflake==0.6.6
certifi==2017.4.17        # via requests
chardet==3.0.4            # via requests
click==6.7                # via python-dotenv
cycler==0.10.0            # via matplotlib
flake8==3.3.0
futures==3.1.1
idna==2.5                 # via requests
isort==4.2.15             # via pylint
jedi==0.10.2
json-rpc==1.8.4
lazy-object-proxy==1.3.1  # via astroid
matplotlib==2.0.2
mccabe==0.6.1             # via flake8, pylint
multidict==3.1.0          # via yarl
numpy==1.13.0             # via matplotlib, pandas
pandas==0.20.2
py==1.4.34                # via pytest
pycodestyle==2.3.1        # via flake8
pyflakes==1.5.0           # via autoflake, flake8
pylint==1.7.1
pyparsing==2.2.0          # via matplotlib
pytest==3.0.7
python-dateutil==2.6.0
python-dotenv==0.6.4
pytz==2017.2              # via matplotlib, pandas
pyyaml==3.12              # via vcrpy
requests==2.18.1
service-factory==0.1.5
six==1.10.0               # via astroid, cycler, matplotlib, pylint, python-dateutil, vcrpy
urllib3==1.21.1           # via requests
vcrpy==1.11.1
wrapt==1.10.10            # via astroid, vcrpy
yapf==0.16.2
yarl==0.11.0              # via vcrpy
